:construction: Problem solve funnel refactoring
